Initial commit
This image is used to have a set of vulnerabilities within a container
environment.

It is using
- https://github.com/rapid7/metasploitable3
and heavily inspired by
- https://github.com/heywoodlh/dockerfiles

The main difference between heywoodlh vulnerable image is that victim
does most of the preparation of metasploitable3 cookbooks upfront
and not while building a image making it a bit easier to be run in a ci.